Refactor Chat Modal to component pieces for easier reuse
